refactor: move ImageData to web extension (#22295)
