+++ ext/canvas/lib.deno_canvas.d.ts
@@ -6,34 +6,6 @@
 /// <reference lib="esnext" />
 
 /** @category Web APIs */
-declare type PredefinedColorSpace = "srgb" | "display-p3";
-
-/** @category Web APIs */
-declare interface ImageDataSettings {
-  readonly colorSpace?: PredefinedColorSpace;
-}
-
-/** @category Web APIs */
-declare interface ImageData {
-  readonly colorSpace: PredefinedColorSpace;
-  readonly data: Uint8ClampedArray;
-  readonly height: number;
-  readonly width: number;
-}
-
-/** @category Web APIs */
-declare var ImageData: {
-  prototype: ImageData;
-  new (sw: number, sh: number, settings?: ImageDataSettings): ImageData;
-  new (
-    data: Uint8ClampedArray,
-    sw: number,
-    sh?: number,
-    settings?: ImageDataSettings,
-  ): ImageData;
-};
-
-/** @category Web APIs */
 declare type ColorSpaceConversion = "default" | "none";
 
 /** @category Web APIs */
