+++ tools/core_import_map.json
@@ -223,6 +223,7 @@
     "ext:deno_web/13_message_port.js": "../ext/web/13_message_port.js",
     "ext:deno_web/14_compression.js": "../ext/web/14_compression.js",
     "ext:deno_web/15_performance.js": "../ext/web/15_performance.js",
+    "ext:deno_web/16_image_data.js": "../ext/web/16_image_data.js",
     "ext:deno_webidl/00_webidl.js": "../ext/webidl/00_webidl.js",
     "ext:deno_websocket/01_websocket.js": "../ext/websocket/01_websocket.js",
     "ext:deno_websocket/02_websocketstream.js": "../ext/websocket/02_websocketstream.js",
